[Refactor] 회원 로그인, 수정 리펙토링
